Reorganised icon generation - Remove circular dependency - Use tempaltes, so it easy to read the code - Use index instead of Icon for Barrel
