+++ scripts/_dummy.template.tsx
@@ -1,0 +1,4 @@
+import * as React from "react"
+// Empty component to make styleguidist happy
+const Dummy: React.FC<{}> = () => <>Dummy</>
+export default Dummy
