+++ src/Breadcrumb/Breadcrumb.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { IconComponentType } from "../Icon/Icon"
+import { IconComponentType } from "../Icon"
 import OperationalContext from "../OperationalContext/OperationalContext"
 import { DefaultProps } from "../types"
 import { inputFocus, isModifiedEvent, isOutsideLink } from "../utils"
