+++ src/Icon/README.md
@@ -36,7 +36,7 @@ import { DocumentIcon } from "@operational/components"
 ```jsx
 import * as React from "react"
 import { Table, ResourceName, Input, SearchIcon, Code, useOperationalContext } from "@operational/components"
-import * as Icon from "@operational/components/Icon/Icon"
+import * as Icon from "@operational/components/Icon"
 
 const MyComponent = () => {
   const [filter, setFilter] = React.useState("")
@@ -73,7 +73,7 @@ const MyComponent = () => {
       </div>
       <Table
         data={Object.keys(Icon)
-          .filter(name => !name.startsWith("_") && name.toLowerCase().includes(filter.toLowerCase()))
+          .filter(name => name.endsWith("Icon") && name.toLowerCase().includes(filter.toLowerCase()))
           .map(name => ({ name }))}
         columns={[
           {
