+++ src/index.ts
@@ -35,7 +35,7 @@ export { default as Group, GroupProps } from "./Group/Group"
 export { default as HeaderBar, HeaderBarProps } from "./HeaderBar/HeaderBar"
 export { default as HeaderMenu, HeaderMenuProps } from "./HeaderMenu/HeaderMenu"
 export { default as Hint, HintProps } from "./Hint/Hint"
-export * from "./Icon/Icon"
+export * from "./Icon"
 export { default as InfoPanel, InfoPanelProps } from "./InfoPanel/InfoPanel"
 export { default as Input, InputProps } from "./Input/Input"
 export { default as Layout, LayoutProps } from "./Layout/Layout"
